% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_roclets_functions.R
\name{roxy_tag_parse.roxy_tag_testexamples}
\alias{roxy_tag_parse.roxy_tag_testexamples}
\title{Parse testexamples tag}
\usage{
\method{roxy_tag_parse}{roxy_tag_testexamples}(x)
}
\arguments{
\item{x}{Input}
}
\value{
Parsed tag
}
\description{
Parse testexamples tag
}
