% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_roclets_functions.R
\name{roxy_tag_parse.roxy_tag_tests}
\alias{roxy_tag_parse.roxy_tag_tests}
\title{Parse tests tag}
\usage{
\method{roxy_tag_parse}{roxy_tag_tests}(x)
}
\arguments{
\item{x}{Input}
}
\description{
Parse tests tag
}
