% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinytest_roclet.R
\name{tinytest_roclet}
\alias{tinytest_roclet}
\title{Roclet: make tinytest test-files.}
\usage{
tinytest_roclet()
}
\description{
This roclet is a workhorse of roxytest,
producing the tinytest test files specified.

Generally you will not call this function directly
but will instead use \code{\link[roxygen2:roxygenise]{roxygen2::roxygenise()}} specifying this roclet.
}
\seealso{
Other roclets:
\code{\link{param_roclet}},
\code{\link{examples_roclet}},
\code{\link{return_roclet}},
\code{\link{testthat_roclet}},
\code{\link[roxygen2]{namespace_roclet}},
\code{\link[roxygen2]{rd_roclet}},
\code{\link[roxygen2]{vignette_roclet}}.

Other roclets: 
\code{\link{examples_roclet}()},
\code{\link{param_roclet}()},
\code{\link{return_roclet}()},
\code{\link{testthat_roclet}()}
}
\concept{roclets}
