% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat_roclet.R
\name{testthat_roclet}
\alias{testthat_roclet}
\title{Roclet: make testthat test-files.}
\usage{
testthat_roclet()
}
\value{
A roclet to be used e.g. with \code{\link[roxygen2:roxygenise]{roxygen2::roxygenise()}}
}
\description{
This roclet is a workhorse of roxytest,
producing the testthat test files specified.

Generally you will not call this function directly
but will instead use \code{\link[roxygen2:roxygenise]{roxygen2::roxygenise()}} specifying this roclet.
}
\examples{
x <- "#' @tests\n#' expect_equal(2, 2)\nNULL\n"
cat(x)
o <- roxygen2::roc_proc_text(testthat_roclet(), x)
cat(o$tests[[1]])

}
\seealso{
Other roclets:
\code{\link{param_roclet}},
\code{\link{examples_roclet}},
\code{\link{return_roclet}},
\code{\link{tinytest_roclet}},
\code{\link[roxygen2]{namespace_roclet}},
\code{\link[roxygen2]{rd_roclet}},
\code{\link[roxygen2]{vignette_roclet}}.

Other roclets: 
\code{\link{examples_roclet}()},
\code{\link{param_roclet}()},
\code{\link{return_roclet}()},
\code{\link{tinytest_roclet}()}
}
\concept{roclets}
