% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat_roclet.R
\name{testthat_roclet}
\alias{testthat_roclet}
\title{Roclet: make testthat test-files.}
\usage{
testthat_roclet()
}
\description{
This roclet is the workhorse of roxytest,
producing the testthat test files specified.

Generally you will not call this function directly
but will instead use roxygenise() specifying the testthat roclet
}
\seealso{
Other roclets:
\code{\link{param_roclet}},
\code{\link[roxygen2]{namespace_roclet}},
\code{\link[roxygen2]{rd_roclet}},
\code{\link[roxygen2]{vignette_roclet}}.

Other roclets: 
\code{\link{param_roclet}()}
}
\concept{roclets}
